added order API

+++ src/main/java/com/example/EatHub/model/Cart.java
@@ -20,7 +20,7 @@ public class Cart {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     int id;
 
-    int cartTotal;
+    double cartTotal;
 
     @OneToOne
     @JoinColumn
